requirements.txt
Kurigram==2.1.39
psutil==7.0.0
pyaes==1.6.1
PySocks==1.7.1
python-dotenv==1.0.1
TgCrypto==1.2.5
